+++ Project/README
@@ -11,7 +11,10 @@ Model : DAO(Data Access Object) DB접근 로직
 
 DTO(Data Transfer Object) : DB접근을 위한 생성자
 	
-쿼리문 **
 
+	
 
-
+	
+- 활용가능한 방법 : 
+View 기능에 대한 메소드들은 interface를 통해 상속받으면 도움이 됨(중복되는 부분이 많음)
+ex) CRUD View에 대한 것들.
