+++ Project/src/controller/ManagerController.java
@@ -2,22 +2,21 @@ package controller;
 
 import java.sql.SQLException;
 
-import dto.InstructorDTO;
 import dto.ManagerDTO;
 import service.ManagerService;
 import view.RunningEndView;
-import view.TestEndView;
 
 public class ManagerController {
 	private static ManagerController instance = new ManagerController();
 	private ManagerService service = ManagerService.getInstance();
-	
-	private ManagerController() {}
-	
+
+	private ManagerController() {
+	}
+
 	public static ManagerController getInstance() {
 		return instance;
 	}
-	
+
 	// 모든 담당자 검색
 	public void allManager() {
 		try {
@@ -26,7 +25,7 @@ public class ManagerController {
 			System.out.println(s);
 		}
 	}
-	
+
 	// 이름으로 특정 담당자 검색
 	public void managerView(String managerName) {
 		try {
@@ -35,7 +34,7 @@ public class ManagerController {
 			e.printStackTrace();
 		}
 	}
-	
+
 	// update/delete를 위한 id값 체크
 	public ManagerDTO checkManagerId(int managerId) {
 		try {
@@ -43,10 +42,10 @@ public class ManagerController {
 		} catch (SQLException e) {
 			e.printStackTrace();
 		}
-		
+
 		return null;
 	}
-	
+
 	// 담당자 정보 수정
 	public void updateManager(ManagerDTO manager, int num, String inputValue) {
 		try {
@@ -55,7 +54,7 @@ public class ManagerController {
 			e.printStackTrace();
 		}
 	}
-	
+
 	// 담당자 정보 추가
 	public void insertManager(ManagerDTO manager) {
 		try {
@@ -64,7 +63,7 @@ public class ManagerController {
 			e.printStackTrace();
 		}
 	}
-	
+
 	// 담당자 정보 삭제
 	public void deleteManager(int managerId) {
 		try {
