--- Project/src/view/TestEndView.java
@@ -1,14 +1,0 @@
-package view;
-
-import dto.InstructorDTO;
-import dto.ManagerDTO;
-
-public class TestEndView {
-	
-	
-	
-	public static void managerView(ManagerDTO manager) {
-		System.out.println(manager);
-	}
- 
-}
