+++ Project/src/view/ViewInterface.java
@@ -1,0 +1,13 @@
+package view;
+
+public interface ViewInterface {
+	public void selectView();
+
+	public void updateView();
+
+	public void allPeople();
+
+	public void insertView();
+
+	public void deleteView();
+}
